+++ tools/ttnn-standalone/ttnn-precompiled.hpp
@@ -19,6 +19,7 @@
 #include "operations/data_movement/repeat/repeat.hpp"
 #include "operations/data_movement/repeat_interleave/repeat_interleave.hpp"
 #include "operations/data_movement/slice/slice.hpp"
+#include "operations/data_movement/sort/sort.hpp"
 #include "operations/data_movement/transpose/transpose.hpp"
 #include "operations/eltwise/binary/binary.hpp"
 #include "operations/eltwise/binary/binary_composite.hpp"
@@ -26,7 +27,6 @@
 #include "operations/eltwise/unary/unary_composite.hpp"
 #include "operations/embedding/embedding.hpp"
 #include "operations/embedding_backward/embedding_backward.hpp"
-#include "operations/experimental/reduction/sort/sort.hpp"
 #include "operations/matmul/matmul.hpp"
 #include "operations/moreh/moreh_cumsum/moreh_cumsum.hpp"
 #include "operations/normalization/batch_norm/batch_norm.hpp"
